+++ web/src/pages/class/create.tsx
@@ -6,7 +6,7 @@ import { v4 } from 'node-uuid'
 import { Redirect, RouteComponentProps } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 
-import { createEditClass, deleteSection, deleteSubject } from 'actions'
+import { createEditClass, deleteSection, deleteSubject, passOutStudents } from 'actions'
 import { AppLayout } from 'components/Layout/appLayout'
 import { PlusButton } from 'components/Button/plus'
 import { blankClass, defaultClasses } from 'constants/form-defaults'
@@ -21,7 +21,10 @@ type State = {
 	subjectToDelete: string
 }
 
-export const CreateOrUpdateClass: React.FC<RouteComponentProps<{ id: string }>> = ({ match }) => {
+export const CreateOrUpdateClass: React.FC<RouteComponentProps<{ id: string }>> = ({
+	match,
+	history
+}) => {
 	const classId = match.params.id
 	const isNewClass = location.pathname.indexOf('new') >= 0
 	const { ref, setIsComponentVisible, isComponentVisible } = useComponentVisible(false)
@@ -150,6 +153,19 @@ export const CreateOrUpdateClass: React.FC<RouteComponentProps<{ id: string }>> 
 
 	if (state.redirectTo) {
 		return <Redirect to={state.redirectTo} />
+	}
+
+	if (!isNewClass && state.class.name === 'Temporary' && state.class.classYear === 9999) {
+		return (
+			<TemporaryClassView
+				passClass={() => {
+					dispatch(passOutStudents(classes[classId]))
+					setTimeout(() => {
+						history.goBack()
+					}, 1000)
+				}}
+			/>
+		)
 	}
 
 	return (
@@ -314,6 +330,57 @@ export const CreateOrUpdateClass: React.FC<RouteComponentProps<{ id: string }>> 
 							</button>
 							<button
 								onClick={() => removeSubject(state.subjectToDelete)}
+								className="py-1 md:py-2 tw-btn-red w-full font-semibold">
+								Confirm
+							</button>
+						</div>
+					</div>
+				</TModal>
+			)}
+		</AppLayout>
+	)
+}
+
+type TemporaryClassViewProps = {
+	passClass: () => void
+}
+
+const TemporaryClassView = ({ passClass }: TemporaryClassViewProps) => {
+	const { ref, setIsComponentVisible, isComponentVisible } = useComponentVisible(false)
+
+	return (
+		<AppLayout title="Finish School" showHeaderTitle>
+			<div className="p-5 md:p-10 md:pt-5 md:pb-0 relative">
+				<div className="md:w-4/5 text-white md:mx-auto flex flex-col items-center space-y-3 rounded-2xl bg-gray-700 py-5 my-4">
+					<h1 className="text-xl md:text-2xl font-medium">Temporary Class</h1>
+					<p className="md:text-center text-lg md:text-xl p-2 leading-relaxed ">
+						This is a class which is automatically generated when you promote students
+						of your final class.
+						<br /> You can delete this class and pass out your students by clicking the{' '}
+						<span className="font-medium italic">Finish School</span> button below
+					</p>
+					<button
+						onClick={() => setIsComponentVisible(true)}
+						className="md:w-1/2 w-11/12 items-center tw-btn-red py-3 font-semibold ">
+						Finish School
+					</button>
+				</div>
+			</div>
+			{isComponentVisible && (
+				<TModal>
+					<div className="bg-white md:p-10 p-8 text-center text-sm" ref={ref}>
+						<div className="font-semibold text-lg">
+							Are you sure you want to proceed? This is an irreversible action!
+						</div>
+
+						<div className="flex flex-row justify-between space-x-4 mt-4">
+							<button
+								onClick={() => setIsComponentVisible(false)}
+								className="py-1 md:py-2 tw-btn bg-gray-400 hover:bg-gray-500 text-white w-full">
+								Cancel
+							</button>
+							<button
+								onClick={() => passClass()}
 								className="py-1 md:py-2 tw-btn-red w-full font-semibold">
 								Confirm
 							</button>