+++ web/CHANGELOG.md
@@ -5,7 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [@mischool-web@v1.0.6] - 2021-07-09
+## [mischool-web@v1.1.0] - 2021-07-15
+
+### Added
+
+-   Promote Students (#384)
+-   Redesigning MIS activation page (#387)
+-   Redesigning student attendance page (#391)
+-   Style tweaks
+-   Link to student profile
+-   Apply max students limit
+
+### Fixes
+
+-   Fix checkPermission() validations
+-
+
+### Fixes
+
+-   Change student and family search input placeholder text
+-   Don't show inactive staff members on staff login page
+
+## [mischool-web@v1.0.6] - 2021-07-09
 
 ### Added